requirements.txt
PySide6
requests
google-generativeai
python-dotenv
coverage
SpeechRecognition 
PyAudio
PyQt5
